Runtime Cell generation
Now the game manager sets up the game state, creates initial cells, properly tracks and sets all relevant tracked cells (eg left and right cell references), deletes "stale" cells when they are outside of the playable area (eg when the monster surpasses that cell), and creates new cells when the player is within a certain distance of the "end" of the area.  Work now will begin on randomizing these cells.