Reworked cell generation code
The randomly generated cells in the experimental branch show some promise but overall a likely equal or greater amount of necessary work to get similar results to the alternative- which is creating a large list of cell prefabs, and giving each of them a list of valid "next" cells.

Work will now begin on creating cell variations.  Later, we will give each of these cell variations their unique list of valid traps.